ccvgd-backend/requirements.txt
apscheduler==3.9.1
Flask==1.1.4
flask_cors==3.0.10
flask_script==2.0.6
Flask_SQLAlchemy==2.5.1
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.27
numpy==1.21.0
pandas==1.3.0
PyMySQL==1.0.2
Werkzeug==1.0.1
MarkupSafe==2.0.1